FIx bug with categorical coloring in plot_cells_3d
When passing a categorical column of colData(cds) as the color_cells_by argument to plot_cells_3d, the color scheme defaults to NA, resulting in no colored points being plotted (just a blank trajectory shows up).

Setting names appropriately to the color palette in the relevant section of code seems to fix the problem.